# runtime/packages/client/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Scrapyomama Runtime API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
# Par défaut : session d'interface (cookie) ou clé d'API (Bearer, scopes de 13 § 8). Les routes publiques posent `security: []` ;
# les routes réservées à la session (registre `auth: 'session'`) posent `security: [{sessionCookie: []}]` : jamais une clé d'API
# sur une route d'administration (13 § 8). Le test de dérive compare ce mode d'authentification au registre du serveur.
security:
  - sessionCookie: []
  - apiKey: []
...
paths:
...
  /api/webhook-subscriptions:
    get:
      operationId: listWebhookSubscriptions
      tags: [webhooks]
      summary: Abonnements webhook (Standard Webhooks, 08 § 5)
      responses:
        '200':
          description: Abonnements.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionList'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      operationId: createWebhookSubscription
      tags: [webhooks]
      summary: Crée un abonnement ; l'URL est contrôlée par la garde SSRF, le secret `whsec_` n'apparaît qu'ici
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookSubscriptionWrite'
      responses:
        '201':
          description: Abonnement créé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionCreated'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /api/webhook-subscriptions/{id}:
    get:
      operationId: getWebhookSubscription
      tags: [webhooks]
      summary: Un abonnement, avec le journal de ses dernières livraisons
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Abonnement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscription'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      operationId: updateWebhookSubscription
      tags: [webhooks]
      summary: Modifie un abonnement (URL, événements, réactivation, rotation du secret)
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookSubscriptionPatch'
      responses:
        '200':
          description: Abonnement à jour ; `secret` présent seulement après une rotation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionCreated'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteWebhookSubscription
      tags: [webhooks]
      summary: Supprime un abonnement
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Supprimé.
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/webhook-subscriptions/{id}/test:
    post:
      operationId: testWebhookSubscription
      tags: [webhooks]
      summary: Bouton **Tester** (06 § 2, Alertes) ; extension de 05 § 4.2
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Résultat du test.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-sy.session
      description: Session d'interface (13 § 5) ; `sy.session` sans le préfixe `__Host-` quand PUBLIC_URL est en http (développement).
    apiKey:
      type: http
      scheme: bearer
      description: Clé d'API `sy_live_…` à scopes (13 § 8) ; jamais de scope d'administration.
    deviceToken:
      type: http
      scheme: bearer
      description: Jeton d'appareil de l'extension (07 § 1), lié à (utilisateur, appareil) ; jamais accepté hors de /api/extension.
    metricsToken:
      type: http
      scheme: bearer
      description: METRICS_TOKEN (14 § 3), distinct des clés d'API.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: "Erreur au format commun `{ error: { code, message } }` (le code est stable, jamais localisé)."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    ReasonCode:
      description: >-
        Code de raison stable (06 § 4.2 : `retried`, `escalated`, `repaired`, `stale`, `reverted`, `cookie_expired`…),
        traduit par la console ; jamais une phrase.
      type: string
      pattern: '^[a-z][a-z0-9_]*$'
    ReasonMessage:
      description: Phrase générée côté serveur, transmise en code et paramètres (traduite par la console, 06 § 4.1).
      type: object
      additionalProperties: false
      required: [code, params]
      properties:
        code:
          $ref: '#/components/schemas/ReasonCode'
        params:
          type: object
          additionalProperties:
            type: [string, number]
...
    WebhookEvent:
      type: string
      enum: [run.succeeded, run.failed, api.status_changed, items.new]
    WebhookDelivery:
      type: object
      additionalProperties: false
      required: [id, at, event, attempt, status_code, duration_ms]
      properties:
        id:
          type: string
        at:
          type: string
          format: date-time
        event:
          description: Événement livré (`WebhookEvent`), ou `webhook.test` pour le bouton Tester.
          type: string
        attempt:
          type: integer
        status_code:
          type: [integer, 'null']
        duration_ms:
          type: [integer, 'null']
        excerpt:
          type: [string, 'null']
    WebhookSubscription:
      type: object
      required: [id, url, events, api_slug, status, tested_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEvent'
        api_slug:
          description: API concernée, ou null pour l'alerte d'instance par défaut.
          type: [string, 'null']
        status:
          type: string
          enum: [active, disabled]
        tested_at:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
        deliveries:
          type: array
          items:
            $ref: '#/components/schemas/WebhookDelivery'
    WebhookSubscriptionWrite:
      type: object
      additionalProperties: false
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          maxLength: 2048
        events:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/WebhookEvent'
        api_slug:
          description: API dont les événements sont livrés ; absent ou null, toutes les API du propriétaire.
          type: [string, 'null']
    WebhookSubscriptionPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        url:
          type: string
          format: uri
          maxLength: 2048
        events:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/WebhookEvent'
        status:
          type: string
          enum: [active, disabled]
        rotate_secret:
          description: Nouveau secret ; l'ancien reste accepté pendant la rotation.
          type: boolean
    WebhookSubscriptionCreated:
      allOf:
        - $ref: '#/components/schemas/WebhookSubscription'
        - type: object
          properties:
            secret:
              description: Secret `whsec_…`, renvoyé une seule fois (création ou rotation).
              type: string
    WebhookSubscriptionList:
      type: object
      additionalProperties: false
      required: [subscriptions]
      properties:
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/WebhookSubscription'
    TestResult:
      type: object
      required: [ok, tested_at]
      properties:
        ok:
          type: boolean
        tested_at:
          type: string
          format: date-time
        error:
          description: Échec lisible, en code stable.
          oneOf:
            - $ref: '#/components/schemas/ReasonMessage'
            - type: 'null'
    # ——— Réglages (08 § 7, 13) : les secrets sont en écriture seule, jamais relus (INV8) ———
...